+++ CHANGELOG.md
@@ -45,6 +45,10 @@
 - Statistics: added 3 new customizable statistics slots based on [X-power mean](https://en.wikipedia.org/wiki/Generalized_mean). Slots and labels can be modified easily at the HTML options panel or at 'Statistics Titleformat Custom-X (p-mean5)' and 'Statistics Titleformat Custom-X (p-mean5)' and 'Statistics Titleformat Custom labels' (properties panel). By default the panel includes custom expressions as example and sets the exponent to 3.
 - Readme: added help submenu, at settings menu, showing all available docs. See [here](https://hydrogenaudio.org/index.php/topic,129076.msg1078914.html#msg1078914).
 - Readme: added Quick help entry, at help submenu, which will show a popup with the most basic functions of the panel. See [here](https://hydrogenaudio.org/index.php/topic,129076.msg1078914.html#msg1078914).
+- External integration: new callbacks.
+	* 'Library-Tree-SMP: cycle show art type'		-> { window: string[], forceShowArt: boolean }
+	* 'Library-Tree-SMP: show art (tf)'				-> { window: string[], forceShowArt: boolean }
+	* 'Library-Tree-SMP: cycle art type'			-> { window: string[], forceShowArt: boolean }
 ### Changed
 - Views: support for $selected{}, $nowplaying{} and $nowplayingorselected{} functions. Views are also updated on real time if such functions are used on playback or selection changes. Note support was only added for display purposes (like highlighting currently playing artist), don't try to use it for filtering purposes; use filters instead. Beware of possible performance impact using this feature. See [here](https://hydrogenaudio.org/index.php/topic,129076.msg1077567.html#msg1077567).
 - Views: updated default view TF patterns with above changes related to $prefix.